requirements.txt
fastapi
uvicorn
qdrant-client
openai
python-dotenv
pydantic